+++ SeasonAffixes/Affixes/Negative/HardWaterAffix.cs
@@ -9,7 +9,7 @@ using SObject = StardewValley.Object;
 
 namespace Shockah.SeasonAffixes;
 
-internal sealed class HardWaterAffix : BaseSeasonAffix, ISeasonAffix
+internal sealed class HardWaterAffix : BaseSeasonAffix, ISeasonAffix // TODO: test in 1.6
 {
 	private static bool IsHarmonySetup = false;